Moved ensemble utilities to quacc.utils.md, updated the defaults (10000 steps, 0.5 fs). We now use ase's built-in function for upper_triangular_cell
