+++ src/quacc/recipes/emt/md.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from ase.calculators.emt import EMT
+from ase.md.npt import NPT
 from ase.md.verlet import VelocityVerlet
 from ase.units import bar, fs
 
@@ -14,6 +15,7 @@ from quacc import Remove, job
 from quacc.runners.ase import Runner
 from quacc.schemas.ase import Summarize
 from quacc.utils.dicts import recursive_dict_merge
+from quacc.utils.md import resolve_md_ensemble, upper_triangular_cell
 
 if TYPE_CHECKING:
     from typing import Any
@@ -21,15 +23,15 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
     from ase.md.md import MolecularDynamics
 
-    from quacc.types import CopyFiles, DynSchema, MDParams
+    from quacc.types import CopyFiles, DynSchema, MDEnsemble, MDParams
 
 
 @job
 def md_job(
     atoms: Atoms,
-    dynamics: MolecularDynamics = VelocityVerlet,
-    steps: int = 1000,
-    timestep_fs: float = 1.0,
+    dynamics: MolecularDynamics | MDEnsemble = VelocityVerlet,
+    steps: int = 10000,
+    timestep_fs: float = 0.5,
     temperature_K: float | None = None,
     pressure_bar: float | None = None,
     md_params: MDParams | None = None,
@@ -45,7 +47,15 @@ def md_job(
     atoms
         Atoms object
     dynamics
-        ASE `MolecularDynamics` class to use, from `ase.md.md.MolecularDynamics`.
+        Either an ASE `MolecularDynamics` class (from `ase.md.md.MolecularDynamics`)
+        or an ensemble name from [quacc.types.MDEnsemble][], resolved to an ASE
+        dynamics class with sensible defaults by
+        [quacc.utils.md.resolve_md_ensemble][] (all of which can be overridden
+        via `md_params["dynamics_kwargs"]`). For the NPT-family ensembles,
+        `pressure_bar` is routed to the appropriate keyword argument of the
+        dynamics class (`externalstress` or `pressure_au`) and defaults to 0 bar
+        when not specified. If the `NPT` class is used, the cell is transformed
+        to the upper-triangular form it requires.
     steps
         Number of MD steps to run.
     timestep_fs
@@ -72,14 +82,21 @@ def md_job(
         Dictionary of results, specified in [quacc.schemas.ase.Summarize.md][].
         See the type-hint for the data structure.
     """
-    md_defaults = {
-        "steps": steps,
-        "dynamics_kwargs": {
-            "timestep": timestep_fs * fs,
+    if isinstance(dynamics, str):
+        dynamics, dynamics_defaults = resolve_md_ensemble(
+            dynamics, timestep_fs * fs, temperature_K, pressure_bar
+        )
+    else:
+        dynamics_defaults = {
             "temperature_K": temperature_K if temperature_K is not None else Remove,
             "pressure_au": pressure_bar * bar if pressure_bar is not None else Remove,
-        },
-    }
+        }
+    dynamics_defaults["timestep"] = timestep_fs * fs
+
+    if dynamics is NPT:
+        atoms = upper_triangular_cell(atoms)
+
+    md_defaults = {"steps": steps, "dynamics_kwargs": dynamics_defaults}
     md_params = recursive_dict_merge(md_defaults, md_params)
 
     calc = EMT(**calc_kwargs)
