+++ tests/core/recipes/mlip_recipes/test_md_recipes.py
@@ -8,18 +8,10 @@ from importlib.util import find_spec
 
 import numpy as np
 from ase.build import bulk
-from ase.md.andersen import Andersen
-from ase.md.bussi import Bussi
 from ase.md.langevin import Langevin
-from ase.md.nose_hoover_chain import MTKNPT, IsotropicMTKNPT, NoseHooverChainNVT
-from ase.md.npt import NPT
-from ase.md.nptberendsen import Inhomogeneous_NPTBerendsen, NPTBerendsen
-from ase.md.nvtberendsen import NVTBerendsen
-from ase.md.verlet import VelocityVerlet
-from ase.units import bar, fs
+from ase.units import fs
 
-from quacc import Remove
-from quacc.recipes.mlip.md import _resolve_ensemble, _upper_triangular_cell, md_job
+from quacc.recipes.mlip.md import md_job
 
 libraries = []
 if has_matcalc := find_spec("matcalc") and find_spec("matgl"):
@@ -65,8 +57,8 @@ def test_md_job(tmp_path, monkeypatch, library):
 
     assert output["name"] == f"{library} MD"
     assert len(output["trajectory"]) == 21
-    assert output["parameters_md"]["timestep"] == pytest.approx(1.0 * fs)
-    assert output["trajectory_log"][10]["time"] == pytest.approx(10 * fs)
+    assert output["parameters_md"]["timestep"] == pytest.approx(0.5 * fs)
+    assert output["trajectory_log"][10]["time"] == pytest.approx(10 * 0.5 * fs)
     assert np.shape(output["results"]["forces"]) == (8, 3)
     # The velocities are seeded, so the initial temperature does not depend on
     # the MLIP library.
@@ -130,7 +122,7 @@ def test_md_job_ensemble_nvt_langevin(tmp_path, monkeypatch, library):
 
     assert output["name"] == f"{library} MD"
     assert len(output["trajectory"]) == 11
-    assert output["parameters_md"]["timestep"] == pytest.approx(1.0 * fs)
+    assert output["parameters_md"]["timestep"] == pytest.approx(0.5 * fs)
     assert output["parameters_md"]["temperature_K"] == 300
     assert output["parameters_md"]["friction"] == pytest.approx(1.0e-3)
     assert output["trajectory_log"][-1]["temperature"] > 0
@@ -167,91 +159,3 @@ def test_md_job_ensemble_npt(tmp_path, monkeypatch, library):
     assert final_cell[1, 0] == final_cell[2, 0] == final_cell[2, 1] == 0.0
     # The input Atoms object should not be mutated.
     assert atoms.cell == pytest.approx(old_cell[:])
-
-
-@pytest.mark.parametrize(
-    ("ensemble", "expected_class"),
-    [
-        ("nve", VelocityVerlet),
-        ("nvt", NoseHooverChainNVT),
-        ("nvt_berendsen", NVTBerendsen),
-        ("nvt_langevin", Langevin),
-        ("nvt_andersen", Andersen),
-        ("nvt_bussi", Bussi),
-        ("npt", NPT),
-        ("npt_berendsen", NPTBerendsen),
-        ("npt_inhomogeneous", Inhomogeneous_NPTBerendsen),
-        ("npt_mtk", MTKNPT),
-        ("npt_isotropic_mtk", IsotropicMTKNPT),
-    ],
-)
-def test_resolve_ensemble_classes(ensemble, expected_class):
-    dynamics, _ = _resolve_ensemble(ensemble, 1.0 * fs, 300, None)
-    assert dynamics is expected_class
-
-
-def test_resolve_ensemble_is_case_insensitive():
-    dynamics, _ = _resolve_ensemble("NVT_Langevin", 1.0 * fs, 300, None)
-    assert dynamics is Langevin
-
-
-def test_resolve_ensemble_coupling_times_scale_with_timestep():
-    _, kwargs = _resolve_ensemble("npt_berendsen", 2.0 * fs, 300, 1.0)
-    assert kwargs["taut"] == pytest.approx(200 * fs)
-    assert kwargs["taup"] == pytest.approx(2000 * fs)
-
-    _, kwargs = _resolve_ensemble("nvt", 2.0 * fs, 300, None)
-    assert kwargs["tdamp"] == pytest.approx(200 * fs)
-
-
-def test_resolve_ensemble_pressure_routing():
-    # ase.md.npt.NPT takes `externalstress`; the Berendsen and MTK families
-    # take `pressure_au`.
-    _, kwargs = _resolve_ensemble("npt", 1.0 * fs, 300, 2.0)
-    assert kwargs["externalstress"] == pytest.approx(2.0 * bar)
-    assert "pressure_au" not in kwargs
-
-    _, kwargs = _resolve_ensemble("npt_berendsen", 1.0 * fs, 300, 2.0)
-    assert kwargs["pressure_au"] == pytest.approx(2.0 * bar)
-    assert "externalstress" not in kwargs
-
-    # The pressure defaults to 0 bar for the NPT-family ensembles.
-    _, kwargs = _resolve_ensemble("npt_mtk", 1.0 * fs, 300, None)
-    assert kwargs["pressure_au"] == 0.0
-
-
-def test_resolve_ensemble_temperature_handling():
-    # An explicit 0 K is honored; only None means unset.
-    _, kwargs = _resolve_ensemble("nvt_langevin", 1.0 * fs, 0, None)
-    assert kwargs["temperature_K"] == 0
-
-    _, kwargs = _resolve_ensemble("nvt_langevin", 1.0 * fs, None, None)
-    assert kwargs["temperature_K"] is Remove
-
-    # NVE ignores the temperature and pressure entirely.
-    _, kwargs = _resolve_ensemble("nve", 1.0 * fs, 300, 1.0)
-    assert kwargs == {}
-
-
-def test_resolve_ensemble_unknown():
-    with pytest.raises(ValueError, match="Unsupported ensemble"):
-        _resolve_ensemble("nvt_gibberish", 1.0 * fs, 300, None)
-
-
-def test_upper_triangular_cell():
-    # The primitive fcc cell is not upper-triangular.
-    atoms = bulk("Cu")
-    old_cell = atoms.cell.copy()
-
-    transformed = _upper_triangular_cell(atoms)
-    new_cell = transformed.cell
-
-    assert transformed is not atoms
-    assert new_cell[1, 0] == new_cell[2, 0] == new_cell[2, 1] == 0.0
-    assert new_cell.cellpar() == pytest.approx(old_cell.cellpar())
-    # The input Atoms object should not be mutated.
-    assert atoms.cell == pytest.approx(old_cell[:])
-
-    # An already upper-triangular cell is returned as-is.
-    cubic = bulk("Cu", cubic=True)
-    assert _upper_triangular_cell(cubic) is cubic
